board hash check
